shared attachmentSummaryComponent
